requirements.txt
app-version==1.0.1
arabic-reshaper==3.0.0
arrow==1.2.1
asgiref==3.7.2
asn1crypto==1.5.1
attrs==21.2.0
Automat==20.2.0
bcc==0.18.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.1.0
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
chrome-gnome-shell==0.0.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
constantly==15.1.0
contourpy==1.2.0
cryptography==42.0.5
cssselect2==0.7.0
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.8
distro==1.7.0
Django==5.0.3
django-appconf==1.0.6
django-auth-ldap==4.8.0
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-permission==1.0.5
django-role-permissions==3.2.0
django-tinymce==4.0.0
djangorestframework==3.15.1
filelock==3.13.1
fonttools==4.50.0
fpdf==1.7.2
fpdf2==2.7.9
greenlet==1.1.2
gunicorn==21.2.0
hidpidaemon==18.4.6
html5lib==1.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
kernelstub==3.1.4
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
ldap3==2.8.1
louis==3.20.0
lxml==5.2.0
macaroonbakery==1.3.1
matplotlib==3.8.3
more-itertools==8.10.0
mysqlclient==1.4.6
netaddr==0.8.0
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
olefile==0.46
oscrypto==1.3.0
packaging==24.0
pbr==5.8.0
pdfkit==1.0.0
pillow==10.3.0
platformdirs==4.2.0
ply==3.11
pop-transition==1.1.2
protobuf==3.12.4
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydbus==0.6.0
pydyf==0.10.0
PyGObject==3.42.1
PyHamcrest==2.0.2
pyHanko==0.23.2
pyhanko-certvalidator==0.26.3
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pypdf==4.1.0
pyphen==0.15.0
pypng==0.20220715.0
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-gnupg==0.4.8
python-ldap==3.4.4
python-pdf==0.40
python-xlib==0.29
pytube==15.0.0
pytz==2022.1
pyxdg==0.27
PyYAML==6.0.1
qrcode==7.4.2
repolib==2.2.1
repoman==1.4.0
reportlab==4.0.9
requests==2.31.0
SecretStorage==3.3.1
service-identity==18.1.0
sessioninstaller==0.0.0
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.31
sqlalchemy-migrate==0.13.0
sqlparse==0.4.4
svglib==1.5.1
systemd-python==234
Tempita==0.5.2
termcolor==2.4.0
tinycss2==1.3.0
Twisted==22.1.0
typing_extensions==4.10.0
tzlocal==5.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
uritools==4.0.2
urllib3==1.26.5
vboxapi==1.0
virtualenv==20.25.1
wadllib==1.3.6
weasyprint==62.2
webencodings==0.5.1
wkhtmltopdf==0.2
WoeUSB-ng==0.2.12
wxPython==4.0.7
xdg==5
xhtml2pdf==0.2.15
xkit==0.0.0
zipp==1.0.0
zope.interface==5.4.0
zopfli==0.2.3
